feat: simpler ticks retrieval
Now we are retrieving the time since boot in ticks instead of adding up
ticks elapsed on all CPUs. This seems to have only one disadvantage:
when running in a VM, this will not account for the time stolen by the
host, so the calculated loads will be in relation to the one physical
core, instead of one virtual core.

At some point we might need a user option to control this behavior. Then
we'll re-add the tick-counting code.