fix: assertion when ppid of some pid is not among pids
In some (extremely rare?) cases, a PID present in /proc could have a
PPID not present in /proc (a race condition perhaps?). This would lead
to the PPID appearing among the keys of the children hashmap failing the
assertion that a sample existed for this key.

The guard against PPID 0 was generalized to all PIDs.